# [spark] Integrate statistics into spark scan