<!-- Please specify the module before the PR name: [core] ... or [flink] ... -->

### Purpose

Integrate statistics into spark scan

### Tests

Test this on 1T tpc-ds, overall performance improved by 15%, some case:

q19 cost from 175s -> 11s
q64 cost from 84s -> 56s
q72 can't run  ->  32s

### API and Format

<!-- Does this change affect API or storage format -->

### Documentation

<!-- Does this change introduce a new feature -->