@YannByron 2024-01-29
https://github.com/apache/incubator-paimon/issues/2404